Corrected mistake on converting user input into floats

+++ rb_101_small_problems/easy_2/exercise_2.rb
@@ -6,7 +6,7 @@ width = gets.chomp
 
 feet_conversion = 10.7639
 
-area_meters = length.to_i * width.to_i
+area_meters = length.to_f * width.to_f
 area_feet = area_meters * feet_conversion
 
 puts "The area of the room is #{area_meters} square meters (#{area_feet} square feet)"
